Bug fix and reorg of quat grad modulus
The modulus of the quaternion gradient included the regularization
needed only for denominator of queternion time evolution.
Also adding new class QuatGradModulusStrategy to do these computations.
